Classes terminées

+++ classes/Item.php
@@ -4,8 +4,11 @@ class Item {
 	public static PDO $db;
 	protected static string $table;
 
-	const SQLMap = [];
+	protected static Array $sqlMap = [];
 
+	public function inDatabase() {
+		return false;
+	}
 	public function toArray() {
 		$arr = [];
 
@@ -18,7 +21,19 @@ class Item {
 	public function forSQL() {
 		$arr = $this->toArray();
 
-		foreach (static::SQLMap as $from => $to) {
+		foreach ($arr as $key => $value) {
+			if ($value instanceof self && !$value->inDatabase()) {
+				throw new InvalidArgumentException(get_class($value) . " " . $key . " is not part of the database");
+			}
+
+			if ($value instanceof DateTime) {
+				$arr[$key] = mysql_timestamp($value);
+			} elseif ($value instanceof self) {
+				$arr[$key] = $value->getDatabaseId();
+			}
+		}
+
+		foreach (static::$sqlMap as $from => $to) {
 			$arr[$to] = $arr[$from];
 			unset($arr[$from]);
 		}
@@ -26,15 +41,24 @@ class Item {
 		return $arr;
 	}
 
-	public static function Find(Array $info) {
-		$request = "SELECT * FROM " . static::$table . " WHERE\n";
-		$i = 0;
-		foreach ($info as $key => $_) {
-			if (is_string($key)) {
-				$request .= "$key = :$key";
-				$i++;
-				if ($i < count($info)) {
-					$request .= " AND \n";
+	public static function Ready() {
+		return self::$db instanceof PDO && is_string(static::$table);
+	}
+
+	public static function Find(Array $info = []) {
+		if (!static::Ready()) return;
+
+		$request = "SELECT * FROM " . static::$table;
+		if (count($info) > 1) {
+			$request .= " WHERE\n";
+			$i = 0;
+			foreach ($info as $key => $_) {
+				if (is_string($key)) {
+					$request .= "$key = :$key";
+					$i++;
+					if ($i < count($info)) {
+						$request .= " AND \n";
+					}
 				}
 			}
 		}
@@ -55,12 +79,14 @@ class Item {
 		$result = static::Find($info);
 		if ($result) $result = $result[0];
 
-		return $result;
+		return $result ?? false;
 	}
 
-	public static function Insert(Item &$item) {
+	public static function Insert(self &$item) {
+		if (!static::Ready()) return;
+
 		if (!$item instanceof static) {
-			return trigger_error(__METHOD__ . " requires an argument " . get_called_class(), E_USER_ERROR);
+			throw new InvalidArgumentException(__METHOD__ . " requires an argument " . get_called_class());
 		}
 
 		$sqlArr = $item->forSQL();
@@ -80,13 +106,15 @@ class Item {
 		}
 	}
 
-	public static function Delete(Item &$item) {
+	public static function Delete(self &$item) {
+		if (!static::Ready()) return;
+
 		if (!$item instanceof static) {
-			return trigger_error(__METHOD__ . " requires an argument " . get_called_class(), E_USER_ERROR);
+			throw new InvalidArgumentException(__METHOD__ . " requires an argument " . get_called_class());
 		}
 
 		if (!$item->getId()) {
-			return trigger_error(__METHOD__ . " requires the " . get_called_class() . " to have an ID assigned", E_USER_ERROR);
+			throw new InvalidArgumentException(__METHOD__ . " requires the " . get_called_class() . " to have an ID assigned");
 		}
 
 		$request = "DELETE FROM " . static::$table . " WHERE id = ?";
@@ -101,13 +129,15 @@ class Item {
 		}
 	}
 
-	public static function Update(Item $item) {
+	public static function Update(self $item) {
+		if (!static::Ready()) return;
+
 		if (!$item instanceof static) {
-			return trigger_error(__METHOD__ . " requires an argument " . get_called_class(), E_USER_ERROR);
+			throw new InvalidArgumentException(__METHOD__ . " requires an argument " . get_called_class());
 		}
 
 		if (!$item->getId()) {
-			return trigger_error(__METHOD__ . " requires the " . get_called_class() . " to have an ID assigned", E_USER_ERROR);
+			throw new InvalidArgumentException(__METHOD__ . " requires the " . get_called_class() . " to have an ID assigned");
 		}
 
 		$sqlArr = $item->forSQL();
